✨ (QRcode) - Restructure qrcode components and add download functionality
- Create new QRCodeCard component to consolidate qrcode display and download options
- Rename printMode to downloadMode in store for clarity
- Remove PrintableView component and integrate its features into new structure
- Add PDF and PNG download capabilities
