+++ package.json
@@ -23,6 +23,7 @@
     "lucide-react": "^0.525.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-error-boundary": "^6.0.0",
     "react-hook-form": "^7.60.0",
     "react-qr-code": "^2.0.18",
     "tailwind-merge": "^3.3.1",
