+++ src/App.tsx
@@ -1,8 +1,5 @@
-import { PrintableView } from "@/components/PrintableView";
-import { QRCodePreview } from "@/components/QRCodePreview";
 import { TravelForm } from "@/components/TravelForm";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { ClipboardList, Printer } from "lucide-react";
+import { QRCodeCard } from "./components/QRCodeCard";
 
 function App() {
   return (
@@ -16,30 +13,10 @@ function App() {
       </header>
 
       <main className="max-w-5xl mx-auto">
-        <Tabs defaultValue="form" className="w-full">
-          <TabsList className="grid w-full grid-cols-2 mb-8">
-            <TabsTrigger value="form" className="flex items-center gap-2">
-              <ClipboardList className="h-4 w-4" />
-              Informations & Aperçu
-            </TabsTrigger>
-            <TabsTrigger value="print" className="flex items-center gap-2">
-              <Printer className="h-4 w-4" />
-              Impression
-            </TabsTrigger>
-          </TabsList>
-
-          <TabsContent
-            value="form"
-            className="flex flex-col md:flex-row gap-6 justify-center"
-          >
-            <TravelForm />
-            <QRCodePreview />
-          </TabsContent>
-
-          <TabsContent value="print" className="flex justify-center">
-            <PrintableView />
-          </TabsContent>
-        </Tabs>
+        <div className="flex flex-col md:flex-row gap-6 justify-center mb-8">
+          <TravelForm />
+          <QRCodeCard />
+        </div>
       </main>
 
       <footer className="max-w-5xl mx-auto mt-12 text-center text-sm text-gray-500">
